Remove deleted hook from Messages.test.js
