+++ src/components/Messages/Messages.js
@@ -3,18 +3,22 @@
 import type { Node } from "react";
 import React from "react";
 
+import searchMessages from "../../lib/searchMessagesAPI";
+import useQuery from "../../sharedHooks/useAuthenticatedQuery";
 import ViewWithFooter from "../SharedComponents/ViewWithFooter";
-import useMessages from "./hooks/useMessages";
 import MessageList from "./MessageList";
 
 const Messages = ( ): Node => {
+  const {
+    data: messages,
+    isLoading
+  } = useQuery( "searchMessages", searchMessages );
   // TODO: Reload when accessing again
-  const { messages, loading } = useMessages( );
 
   return (
     <ViewWithFooter>
       <MessageList
-        loading={loading}
+        loading={isLoading}
         messageList={messages}
         testID="Messages.messages"
       />
